some restructuring of how to handle asynch events. The client now make some commands synchronous over asynchronous calls to the server. This applies only to commands where the client can "figure out" when a response has come back.
E.g: If a login is happening then the calling process is blocking while the client process waits for either the welcome message or an error. Meanwhile it is not expected that there is any other calling process doing any calls

E.g2: If quit/2 is used then the client expects to receive a response from the server saying it is closing the link. It is not expected that the calling process will send a quit command and then do something else so it is blocking.

Note: It is _strongly_ assumed that there is only _one_ calling process!