Changed the handling of CTCP events somewhat. There is now a flag in the #ircmsg record called ctcp. It can have one of the following values: true, false or invalid.
true means that the message was interpreted as an CTCP message (as specified informally on the internet and conforms with how mIRC sends/receives CTCP messages).

false means it is anything other then a CTCP message.

invalid means that it looked as if it was an CTCP message but it turned out that it wasn't conformant with the proper format. A CTCPMsg must be wrapped inside a PRIVMSG command and the second argument must be the following format: CTCPMsg = "'\001'<Chars-Except-001-Char>'\001'" Thus the whole thing looks like this when coming from the server: "PRIVMSG <nick> :<CTCPMsg>\r\n", If it doesn't then the ctcp flag will be invalid. Invalid CTCP messages will only be delivered back to the controlling process if the "{debug, true}" option has been passed with the initial Options when starting the client.