Add supervisor to monitor the client and extend some APIs
The client is started under a supervisor, this lets the user run
multiple clients which are supervised without taking care of the code
for it. One can still use the eirc_cl:x module to start the clients
directly and bypass the supervisor if the functionality is not needed.

Some APIs for introspection were added and the idea now is that a "bot"
must register with the client to get events, it is no longer assumed
that the process that starts the client is the one that wants to receive
the irc messages. To connect and login easier the logon API was given an
easier API for one that is not familiar with the IRC protocol and
usually (can't think of when not) a user wants to logon after connecting
so this was moved into one API.