Add on_ping, on_kick, on_nick and on_raw callbacks
on_ping happens when the server sends a ping request. on_kick happens
when someone is kicked from a channel the bot is in. on_nick happens
when someone changes his/her nickname in a channel and on_raw is sent if
none of the previous commands are picked up by the behaviour and gives
the bot a chance to respond to them.